% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{abs,sparse_numeric-method}
\alias{abs,sparse_numeric-method}
\title{Absolute Value Method}
\usage{
\S4method{abs}{sparse_numeric}(x)
}
\arguments{
\item{x}{A sparse_numeric object}
}
\description{
Absolute Value Method
}
