% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{norm,sparse_numeric-method}
\alias{norm,sparse_numeric-method}
\title{Euclidean Norm Method}
\usage{
\S4method{norm}{sparse_numeric}(x, ...)
}
\arguments{
\item{x}{A sparse_numeric object}

\item{...}{Ignored}
}
\description{
Euclidean Norm Method
}
