% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{norm}
\alias{norm}
\title{Generic for sparse Euclidean norm}
\usage{
norm(x, ...)
}
\arguments{
\item{x}{A sparse_numeric object.}

\item{...}{Ignored.}
}
\description{
Generic for sparse Euclidean norm
}
