% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sn_vec}
\alias{sn_vec}
\title{Construct a sparse_numeric object}
\usage{
sn_vec(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A sparse_numeric object.
}
\description{
Construct a sparse_numeric object
}
