% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sparse_add,sparse_numeric,sparse_numeric-method}
\alias{sparse_add,sparse_numeric,sparse_numeric-method}
\title{Sparse Addition Method}
\usage{
\S4method{sparse_add}{sparse_numeric,sparse_numeric}(x, y, ...)
}
\arguments{
\item{x}{A sparse_numeric object}

\item{y}{A sparse_numeric object}

\item{...}{Ignored}
}
\description{
Sparse Addition Method
}
