% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sparse_arithmetic}
\alias{sparse_arithmetic}
\alias{sparse_add}
\alias{sparse_add,sparse_numeric,sparse_numeric-method}
\alias{+,sparse_numeric,sparse_numeric-method}
\alias{sparse_mult}
\alias{sparse_mult,sparse_numeric,sparse_numeric-method}
\alias{*,sparse_numeric,sparse_numeric-method}
\alias{sparse_sub}
\alias{sparse_sub,sparse_numeric,sparse_numeric-method}
\alias{-,sparse_numeric,sparse_numeric-method}
\title{Arithmetic operations for sparse_numeric}
\usage{
sparse_add(x, y, ...)

\S4method{sparse_add}{sparse_numeric,sparse_numeric}(x, y, ...)

\S4method{+}{sparse_numeric,sparse_numeric}(e1, e2)

sparse_mult(x, y, ...)

\S4method{sparse_mult}{sparse_numeric,sparse_numeric}(x, y, ...)

\S4method{*}{sparse_numeric,sparse_numeric}(e1, e2)

sparse_sub(x, y, ...)

\S4method{sparse_sub}{sparse_numeric,sparse_numeric}(x, y, ...)

\S4method{-}{sparse_numeric,sparse_numeric}(e1, e2)
}
\arguments{
\item{x}{A sparse_numeric object}

\item{y}{A sparse_numeric object}

\item{...}{Additional arguments (currently unused)}

\item{e1}{A sparse_numeric object (used for S4 +, -, * methods)}

\item{e2}{A sparse_numeric object (used for S4 +, -, * methods)}
}
\description{
S4 methods for \code{+}, \code{-}, and \code{*} applied to sparse_numeric objects.
}
