% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sparse_crossprod}
\alias{sparse_crossprod}
\title{Generic for sparse crossproduct}
\usage{
sparse_crossprod(x, y, ...)
}
\arguments{
\item{x}{A sparse_numeric object.}

\item{y}{A sparse_numeric object.}

\item{...}{Ignored.}
}
\description{
Generic for sparse crossproduct
}
