% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{-,sparse_numeric,sparse_numeric-method}
\alias{-,sparse_numeric,sparse_numeric-method}
\title{Sparse Subtraction Operator}
\usage{
\S4method{-}{sparse_numeric,sparse_numeric}(e1, e2)
}
\arguments{
\item{e1}{A sparse_numeric object}

\item{e2}{A sparse_numeric object}
}
\description{
Sparse Subtraction Operator
}
