% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sparse_sub}
\alias{sparse_sub}
\title{Generic for sparse subtraction}
\usage{
sparse_sub(x, y, ...)
}
\arguments{
\item{x}{A sparse_numeric object.}

\item{y}{A sparse_numeric object.}

\item{...}{Ignored.}
}
\description{
Generic for sparse subtraction
}
