% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{standardize}
\alias{standardize}
\title{Generic for standardizing sparse vectors}
\usage{
standardize(x, ...)
}
\arguments{
\item{x}{A sparse_numeric object.}

\item{...}{Ignored.}
}
\description{
Generic for standardizing sparse vectors
}
