% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\docType{class}
\name{sparse_numeric-class}
\alias{sparse_numeric-class}
\alias{show,sparse_numeric-method}
\title{sparse_numeric S4 Class}
\usage{
\S4method{show}{sparse_numeric}(object)
}
\arguments{
\item{object}{A sparse_numeric object}
}
\description{
Stores a numeric vector in sparse form using nonzero values and positions.
}
\section{Slots}{

\describe{
\item{\code{value}}{Numeric vector of non-zero values.}

\item{\code{pos}}{Integer vector of positions (1-based indices).}

\item{\code{length}}{Integer total length of the vector.}
}}

